+++ src/main/java/de/tum/bgu/msm/moped/modules/walkModeChoice/HBCollegeWalk.java
@@ -12,9 +12,9 @@ public class HBCollegeWalk {
 
     private static final Logger logger = Logger.getLogger(HBCollegeWalk.class);
     private final DataSet dataSet;
-    private Table<Integer, Integer, Double> WalkTrip;
-    private Table<Integer, Integer, Double> WalkExpUtility;
-    private Table<Integer, Integer, Double> VehicleExpUtility;
+    private Table<Long, Integer, Double> WalkTrip;
+    private Table<Long, Integer, Double> WalkExpUtility;
+    private Table<Long, Integer, Double> VehicleExpUtility;
 
 
 
@@ -23,7 +23,7 @@ public class HBCollegeWalk {
     }
 
     public void run () {
-        Collection<Integer> zones = dataSet.getZones().keySet();
+        Collection<Long> zones = dataSet.getZones().keySet();
         Collection<Integer> households = dataSet.getHhTypes().keySet();
         WalkTrip = ArrayTable.create(zones, households);
         WalkExpUtility = ArrayTable.create(zones, households);
@@ -35,7 +35,7 @@ public class HBCollegeWalk {
     }
 
     private void walkUtilityCalculator() {
-        for (int zoneId : dataSet.getZones().keySet()) {
+        for (long zoneId : dataSet.getZones().keySet()) {
             //zonal-related utility
             double stfwy = dataSet.getZone(zoneId).getStfwy();
             double pie = dataSet.getZone(zoneId).getPie();
@@ -59,7 +59,7 @@ public class HBCollegeWalk {
     }
 
     private void vehicleUtilityCalculator() {
-        for (int zoneId : dataSet.getZones().keySet()) {
+        for (long zoneId : dataSet.getZones().keySet()) {
             //TODO: zonal-related utility
             double utilityZone = 0.0;
 
@@ -74,7 +74,7 @@ public class HBCollegeWalk {
     }
 
     private void walkTripsCalculator() {
-        for (int zoneId : dataSet.getZones().keySet()) {
+        for (long zoneId : dataSet.getZones().keySet()) {
             for (int hhTypeId : dataSet.getHhTypes().keySet()) {
                 double walkExpUtility = WalkExpUtility.get(zoneId,hhTypeId);
                 double vehicleExpUtility = VehicleExpUtility.get(zoneId,hhTypeId);
